+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 3.0 Changes on 2021-03
+## 3.1 Changes on 2021-
+- Update Api-Version to 6.3
+- Allow WebPaymentReference to be set on PaymentModel
+- Add AuthCode to initial receipt response
+- Add AuthCode, WebPaymentReference and Acquirer to response of Transactions.Get(receiptId)
+
+## 3.0 Changes on 2021-03-11
 - Update default base Urls with certificate pinning checks
 - Update Api-Version to 6.2
 - Add Address1/Address2/Address3 to CardAddressModel (deprecate Line1/Line2/Line3)